worked on Race
Added BaseEntity, a MappedSuperclass for tables that have the common name and description tables. Extended BaseEntity in the Race class. Added a POST request to RaceController & RaceService. Fixed the annotations for the id field in Race to let the database handle sequencing the id column. Added some validation logic to the new saveRace POST method and the fail fast pattern.