moved raceAlreadyExists to RepositoryHelper
I'm going to be checking to see if the name field already exists quite a fair bit across the services, so I moved the raceAlreadyExists() method to the RepositoryHelper and renamed the method to be more descriptive. I also renamed the findRaceByName() method in RaceRepository to a more generic name in order to increase reusability of the new nameAlreadyExists() method's helper method (getRepoMethod()). I hope that makes sense when/if I re-read this.