added ForeignKey enum
I didn't want to have a getForeignKeyColumnName() method in every relevant service so I decided to move the fk column names to an enum and reference it where I need it. Updated the tests and Continent to reference the new enum where I was explicitly passing in the column name before.