created campaigns, users, and permissions tables, added campaign_uuid as fk to tables where users will have crud rights
I'm not 100% sure on what needs to go into the Permissions table. The idea is that Campaigns is the highest level and all the other tables are associated with it. Users can be associated with campaigns. The Permissions table will help control which user has CRUD rights to which campaign and will provide enough flexibility for co-DMing/co-GMing situations and/or swapping DMs mid-game (hopefully). The fk_campaign_uuid column is to tie data to a specific campaign. (Bob should be able to see the stuff he's input for his campaign, but not Jane's campaign info and vice versa)